ZOODATA_VISION/requirements.txt
opencv-python
numpy
matplotlib
pandas
megadetector
scikit-learn
scikit-image
umap-learn
timm
hdbscan
streamlit